% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{dl_tai_table}
\alias{dl_tai_table}
\title{Download TAI data table}
\usage{
dl_tai_table(variable, vars = NULL, lang = c("et", "en"))
}
\arguments{
\item{variable}{a table Name from list_variables(), a character string.}

\item{vars}{table of available variables, a data_frame generated by
list_variables() function. Defaults to NULL, in which case internal
table is used.}

\item{lang}{Language, a character string. "et" -- Estonian, "en" -- English.}
}
\description{
Download TAI data table
}
\examples{
# Download data for table 'RK01'
rk01 <- dl_tai_table("RK01", lang = "en")
rk01

}
