% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_table.R
\name{download_table}
\alias{download_table}
\title{Download variable values}
\usage{
download_table(variable, vars = list_variables())
}
\arguments{
\item{variable}{variable name, a character string.}

\item{vars}{table of available variables, a data_frame generated by
list_variables() function. Defaults to NULL, in which case internal
table is used.}
}
\description{
Download variable values
}
