% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_table.R
\name{eliminate}
\alias{eliminate}
\title{Check vars that can be eliminated}
\usage{
eliminate(x)
}
\arguments{
\item{x}{metadata from pxweb database, a data frame.}
}
\description{
Check vars that can be eliminated
}
