% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_icd.R
\name{extract_icd}
\alias{extract_icd}
\title{Extract ICD10 string.
Extracts ICD10 string and expands range when range of codes is shown.}
\usage{
extract_icd(x)
}
\arguments{
\item{x}{character string containing ICD10 code.}
}
\value{
Character vector with ICD10 codes.
}
\description{
Extract ICD10 string.
Extracts ICD10 string and expands range when range of codes is shown.
}
\examples{
icd <- extract_icd("All malignant neoplasms (C00-C97)")
}
