% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_api.R
\name{get_available_databases}
\alias{get_available_databases}
\title{List of available databases at TAI, wrapper around tai_api}
\usage{
get_available_databases()
}
\value{
content a data.frame of database ids and title

path the path used in API query

response html response
}
\description{
List of available databases at TAI, wrapper around tai_api
}
\examples{
db <- get_available_databases()
db$content
}
