% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_api.R
\name{get_json}
\alias{get_json}
\title{Extract json as data frame from request content}
\usage{
get_json(resp)
}
\arguments{
\item{resp}{request object.}
}
\value{
a data_frame
}
\description{
Checks if request object type is \code{application/json}, retrieves contents and converts from json to data frame.
}
