% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_api.R
\name{get_nodes}
\alias{get_nodes}
\title{Get database nodes}
\usage{
get_nodes(dbi)
}
\arguments{
\item{dbi}{TAI database id}
}
\value{
a data.frame with id, type - types of node object l and t where l is s sublevel and t is a table, text - textual description
}
\description{
Get database nodes
}
