% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{icd_intersect}
\alias{icd_intersect}
\title{Evaluates whether intersect between two vectors with ICD10 codes is longer than zero.}
\usage{
icd_intersect(x)
}
\arguments{
\item{x}{List of length two. Two character vectors with ICD10 codes.}
}
\value{
logical.
}
\description{
Evaluates whether intersect between two vectors with ICD10 codes is longer than zero.
}
