% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_var.R
\name{is_summary}
\alias{is_summary}
\title{Detect if variable is a summary variable. Summary variable name has less
dots in the prefix than immediately following variable.}
\usage{
is_summary(x)
}
\arguments{
\item{x}{Integer vector. Number of dots prefixed to variable name.}
}
\value{
Logical vector. Summary variables are TRUE.
}
\description{
Detect if variable is a summary variable. Summary variable name has less
dots in the prefix than immediately following variable.
}
