% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_variables.R
\name{list_variables}
\alias{list_variables}
\title{List available variables}
\usage{
list_variables(local = TRUE, lang = c("et", "en"), verbose = FALSE)
}
\arguments{
\item{local}{whether to return package internal table of available datasets. A logical, defaults to TRUE.}

\item{lang}{Language. Estonian 'et' or english 'en'. Affects table titles and variable names.}

\item{verbose}{Display messages to track query progress, defaults to TRUE.}
}
\value{
A data_frame with following columns:
\describe{
  \item{Database}{Database name as available at TAI.}
  \item{Node}{Database node name.}
  \item{Name}{Variable name. This name should be used to download variable values.}
  \item{Title}{Description of the variable.}
  \item{Updated}{Time of the last update.}
}
}
\description{
Lists available databases, nodes and variables in TAI/National Institute for Health Development database.
}
\examples{
\dontrun{
vars <- list_variables()

# In RStudio, it is convenient to use View function
View(vars)
}

}
