% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_api.R
\name{query_api}
\alias{query_api}
\title{Run query using GET}
\usage{
query_api(path = NULL, url)
}
\arguments{
\item{path}{optional path appended to url, defaults to NULL.}

\item{url}{the url of the page to retrieve.}
}
\description{
Wrapper around httr::GET
}
