% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_table.R
\name{query_fun}
\alias{query_fun}
\title{Run api query}
\usage{
query_fun(x, sleep)
}
\arguments{
\item{x}{a table with database and node ids}

\item{sleep}{sleep time between queries in seconds to avoid server choking}
}
\description{
Run api query
}
