% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_var.R
\name{summary_var}
\alias{summary_var}
\title{Identify summary variables using dots prefixes in variable names.}
\usage{
summary_var(var)
}
\arguments{
\item{var}{Character or factor vector. Variable names from from Estonian
Health Statistics and Health Research database.}
}
\description{
Identify summary variables using dots prefixes in variable names.
}
\examples{
# "K천ik p천hjused (A00-Y89)" includes all following variables and
# "Nakkus- ja parasiithaigused (A00-B99)" includes all following variables,
# prefixed with "..".
vars <- c("K천ik p천hjused (A00-Y89)",  "Nakkus- ja parasiithaigused (A00-B99)",
         "..tuberkuloos (A15-A19, B90)", "..meningokokknakkus (A39)",
         "..sepsis (A40-A41)", "..viirushepatiit (B15-B19)")
sumvars <- summary_var(vars)
}
