% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_api.R
\name{tai_api}
\alias{tai_api}
\title{Interact with TAI API.}
\usage{
tai_api(path)
}
\arguments{
\item{path}{API path relative to http://pxweb.tai.ee}
}
\value{
Returns API response object with three slots:
\itemize{
\item content parsed query result
\item path query path
\item response query response
}
}
\description{
Interact with TAI API.
}
