% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_table.R
\name{traverse_nodes}
\alias{traverse_nodes}
\title{Run query on database nodes}
\usage{
traverse_nodes(node, sleep)
}
\arguments{
\item{node}{node id, character string}

\item{sleep}{time between querying batches}
}
\description{
Run query on database nodes
}
