% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_json.R
\name{json_to_df}
\alias{json_to_df}
\title{Extract data from json file downloaded from Estonian Health Statistics database}
\usage{
json_to_df(x)
}
\arguments{
\item{x}{A complex json file.}
}
\value{
Returns data.frame in the long format.
}
\description{
Extract data from json file downloaded from Estonian Health Statistics database
}
\examples{
system.file("extdata", "PK10.json", package = "boulder")
library(jsonlite)
pk10js <- fromJSON("PK10.json")
pk10 <- json_to_df(pk10js)

}
