% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_table.R
\name{pull_table}
\alias{pull_table}
\title{Download data table from TAI}
\usage{
pull_table(tabname, tablist = NULL, lang = c("et", "en"))
}
\arguments{
\item{tabname}{a table Name from list_variables(), a character string.}

\item{tablist}{list of available database tables, a data_frame generated by
get_all_tables() function. Defaults to NULL, in which case internal
table is used.}

\item{lang}{Language, a character string. "et" -- Estonian, "en" -- English.}
}
\description{
Downloads data table in Estonian National Institute for Health Development (Tervise Arengu Instituut, TAI) database.
}
\examples{
# Download data for table 'RK01'
rk01 <- pull_table("RK01", lang = "en")
rk01

}
