% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_table.R
\name{recode_vars}
\alias{recode_vars}
\title{Recode downloaded variables}
\usage{
recode_vars(vector_to_modify, replacements)
}
\arguments{
\item{vector_to_modify}{A vector to modify.}

\item{replacements}{Replacements.}
}
\description{
Recode downloaded variables
}
