% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_table.R
\name{recode_vars}
\alias{recode_vars}
\title{Recode downloaded variables}
\usage{
recode_vars(var, lookup)
}
\arguments{
\item{var}{original variable, atomic vector.}

\item{lookup}{named list with original variable and replacement.}
}
\description{
Recode downloaded variables
}
