% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tai_api.R
\name{tai_api}
\alias{tai_api}
\title{#' Interact with TAI API.}
\usage{
tai_api(path)
}
\arguments{
\item{path}{API path relative to http://pxweb.tai.ee}

\item{tidy}{convert response into data_frame}
}
\value{
API response object with three slots:
- content parsed query result
- path query path
- response query response
}
\description{
#' Interact with TAI API.
}
\examples{
# Return available databases
path <- "/PXWeb2015/api/v1/et"
db <- tai_api(path)
db
}
