% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_table.R
\name{unwrap_content}
\alias{unwrap_content}
\title{Extract table from downloaded content.}
\usage{
unwrap_content(tab)
}
\arguments{
\item{tab}{a table to be unwrapped.}
}
\description{
Extract table from downloaded content.
}
